requirements-dev.txt
black==21.9b0
eth_brownie==1.17
gnosis-py==3.6.0
ape_safe==0.5.0
click==8.0.3
regex==2021.10.8
tomli==1.2.1
tenacity==8.0.1
psutil==5.8.0 
python-dotenv==0.16.0